requirements.txt
confluent-kafka
requests
six
pandas
pymongo[srv]==4.1.1
dnspython
-e .

pandas
notebook
python-dotenv==1.0.1